+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog].
 - Small performance improve `#641`
 - Ability to prevent changing enablement of component `#668`
 - Remove Zero Sized Polygons `#659`
+- Add support for UniVRM components `#653`
 
 ### Changed
 - All logs passed to ErrorReport is now shown on the console log `#643`